Update tests to be port independent
Some test cases start a new server
on Port:3000, which results in EADDRINUSE
failure on CI when port:3000 is busy.
This commit addresses above issue by setting
port to 0 which lets the OS to allocate any
avilable port, dynamically, to be used for
starting the server, which makes test cases
more robust.